Improve the API documentation
