+++ Cargo.toml
@@ -144,7 +144,7 @@ name = "query"
 name = "raw"
 
 [package.metadata.docs.rs]
-features = ["bundled", "complete", "serialized"]
+features = ["bundled", "derive", "complete", "serialized"]
 no-default-features = true
 
 rustdoc-args = ["--cfg", "docsrs"]
